refactor(cli): extract review subcommand into cli/commands/review.mjs
Move the `review` (plan|exec|verify|route) block out of main() into a
dedicated handler. Behavior, messages, and exit codes are unchanged; only
the enclosing function and relative import depth differ.
